pyproject.toml
[project]
name = "pallets-web"
version = "1.0.0"
description = "Pallets website and content."
readme = "README.md"
license = {file = "LICENSE.txt"}
maintainers = [{name = "Pallets", email = "contact@palletsprojects.com"}]
classifiers = ["Private :: No Not Upload"]
requires-python = "~=3.12.0"
dependencies = [
    "feedgen",
    "flask",
    "flask-sqlalchemy",
    "gunicorn",
    "markdown-it-py",
    "pygments",
]

[project.urls]
Home = "https://palletsprojects.com"
Source = "https://github.com/pallets/pallets-web/"

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "pallets"

[tool.mypy]
files = ["src/pallets"]
show_error_codes = true
pretty = true
strict = true

[tool.pyright]
include = ["src/pallets"]

[tool.ruff]
src = ["src"]
fix = true
unsafe-fixes = true
show-fixes = true
output-format = "full"

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # pyflakes
    "I",  # isort
    "UP",  # pyupgrade
    "W",  # pycodestyle warning
]
ignore-init-module-imports = true

[tool.ruff.lint.isort]
force-single-line = true
order-by-type = false
